# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Monitoring platform 1000 Moustaches API",
    "description": "API documentation for the 1000 Moustaches monitoring platform",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/veterinarians/": {
      "get": {
        "tags": [
          "Veterinarians"
        ],
        "summary": "Get all veterinarians",
        "description": "Retrieve a list of all veterinarians in the system",
        "responses": {
          "200": {
            "description": "List of veterinarians",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Veterinarian"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Veterinarians"
        ],
        "summary": "Create new veterinarian",
        "description": "Register a new veterinarian in the system",
        "responses": {
          "201": {
            "description": "Veterinarian created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veterinarian"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Veterinarian"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVeterinarianRequest"
              }
            }
          }
        }
      }
    },
    "/veterinarians/{id}": {
      "get": {
        "tags": [
          "Veterinarians"
        ],
        "summary": "Get veterinarian by ID",
        "description": "Retrieve a specific veterinarian by their ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Veterinarian ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Veterinarian details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veterinarian"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Veterinarian"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Veterinarians"
        ],
        "summary": "Update veterinarian",
        "description": "Update an existing veterinarian\\'s information",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Veterinarian ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Veterinarian updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veterinarian"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Veterinarian"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVeterinarianRequest"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Veterinarians"
        ],
        "summary": "Delete veterinarian",
        "description": "Remove a veterinarian from the system",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Veterinarian ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Veterinarian deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/veterinarian-interventions/": {
      "get": {
        "tags": [
          "Veterinary Interventions"
        ],
        "summary": "Get all veterinary interventions",
        "description": "Retrieve a list of all veterinary interventions in the system",
        "responses": {
          "200": {
            "description": "List of veterinary interventions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VeterinarianIntervention"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Veterinary Interventions"
        ],
        "summary": "Create new veterinary intervention",
        "description": "Record a new veterinary intervention for an animal",
        "responses": {
          "201": {
            "description": "Veterinary intervention created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VeterinarianIntervention"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/VeterinarianIntervention"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVeterinarianInterventionRequest"
              }
            }
          }
        }
      }
    },
    "/veterinarian-interventions/{id}": {
      "get": {
        "tags": [
          "Veterinary Interventions"
        ],
        "summary": "Get veterinary intervention by ID",
        "description": "Retrieve a specific veterinary intervention by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Veterinary intervention ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Veterinary intervention details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VeterinarianIntervention"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/VeterinarianIntervention"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Veterinary Interventions"
        ],
        "summary": "Update veterinary intervention",
        "description": "Update an existing veterinary intervention",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Veterinary intervention ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Veterinary intervention updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VeterinarianIntervention"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/VeterinarianIntervention"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVeterinarianInterventionRequest"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Veterinary Interventions"
        ],
        "summary": "Delete veterinary intervention",
        "description": "Remove a veterinary intervention record",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Veterinary intervention ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Veterinary intervention deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Authorization": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Firebase authentication token"
      }
    },
    "schemas": {
      "Species": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Species name",
            "example": "Chat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
...
      "Animal": {
        "type": "object",
        "required": [
          "name",
          "icad"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Animal name",
            "example": "Whiskers"
          },
          "species": {
            "$ref": "#/components/schemas/Species"
          },
          "icad": {
            "type": "string",
            "description": "ICAD identification number",
            "example": "250268500123456"
          },
          "sexe": {
            "type": "string",
            "description": "Animal gender",
            "enum": [
              "M",
              "F"
            ],
            "example": "F"
          },
          "race": {
            "type": "string",
            "description": "Animal breed",
            "example": "European Shorthair"
          },
          "adopted": {
            "type": "boolean",
            "description": "Is adopted",
            "default": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
...
      "Veterinarian": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Veterinarian name",
            "example": "Dr. Martin"
          },
          "phone": {
            "type": "string",
            "description": "Phone number",
            "example": "+33123456789"
          },
          "mail": {
            "type": "string",
            "format": "email",
            "description": "Email address",
            "example": "dr.martin@veterinary.com"
          },
          "address": {
            "type": "string",
            "description": "Veterinarian address",
            "example": "123 Rue de la Paix, 75001 Paris"
          },
          "priceLevel": {
            "type": "integer",
            "description": "Price level (1-5 scale)",
            "minimum": 1,
            "maximum": 5,
            "example": 3
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
      "VeterinarianIntervention": {
        "type": "object",
        "required": [
          "date"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier",
            "example": 1
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Date when intervention was performed",
            "example": "2023-12-15T10:30:00Z"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the intervention"
          },
          "animal": {
            "$ref": "#/components/schemas/Animal"
          },
          "veterinarian": {
            "$ref": "#/components/schemas/Veterinarian"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
...
      "CreateVeterinarianRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Dr. Martin"
          },
          "phone": {
            "type": "string",
            "example": "+33123456789"
          },
          "mail": {
            "type": "string",
            "format": "email",
            "example": "dr.martin@veterinary.com"
          },
          "address": {
            "type": "string",
            "example": "123 Rue de la Paix, 75001 Paris"
          },
          "priceLevel": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "example": 3
          }
        }
      },
      "CreateVeterinarianInterventionRequest": {
        "type": "object",
        "required": [
          "date"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Date when intervention was performed",
            "example": "2023-12-15T10:30:00Z"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the intervention"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "Authorization": []
    }
  ]
}